Set up outline for using posix_spawn rather than system(),
I looked up system functions for copy, but those seem to be unique to
C++17.  Removing a file might be done with unlink(), but I don't know
the C API that well, so copying what I did for posix_spawn() it is.

+++ tools/registry.c
@@ -157,7 +157,23 @@ main( int argc, char *argv[] )
       sprintf( fname_wrk,"%s/Registry_irr_diag",dir ) ;
     }
 //  fprintf(stderr,"Registry tmp file = %s\n",fname_wrk);
-    sprintf(command,"/bin/cp %s %s\n",fname_in,fname_wrk);
+    /* we should be able to implement this using posix_spawn */
+    /*
+      #include <spawn.h>
+      extern char **environ;
+      pid_t child_pid;
+      // There doesn't seem to be a way to specify the length, so I'm
+      // assuming it's null-terminated
+      char *command_argv[4] = {"/bin/cp", NULL, NULL, NULL};
+      command_argv[1] = fname_in;
+      command_argv[2] = fname_wrk;
+
+      if (posix_spawn(&child_pid, command_argv[0], NULL, NULL, command_argv, environ)) {
+        fprintf(stderr, "Could not copy %s to %s\n", fname_in, fname_wrk);
+	exit(2);
+      }
+     */
+    sprintf(command,"/bin/cp \'%s\' \'%s\'\n",fname_in,fname_wrk);
 //  fprintf(stderr,"Command = %s\n",command);
     if( system( command ) ) {
       fprintf(stderr,"Could not copy %s to %s\n",fname_in,fname_wrk);
@@ -169,6 +185,15 @@ main( int argc, char *argv[] )
       exit(2) ;
     }
     if( !access( "Registry/registry.irr_diag",F_OK ) ) {
+      /*
+	command_argv[0] = "/bin/rm";
+	command_argv[1] = "-f";
+	command_argv[2] = "Registry/registry.irr_diag";
+	if (posix_spawn(&child_pid, command_argv[0], NULL, NULL, command_argv, environ)) {
+          fprintf(stderr, "Could not remove Registry/registry.irr_diag\n", fname_in, fname_wrk);
+	  exit(2);
+	}
+      */
       sprintf(command,"/bin/rm -f Registry/registry.irr_diag\n");
       if( system( command ) ) {
         fprintf(stderr,"Could not remove Registry/registry.irr_diag\n");
@@ -279,6 +304,15 @@ cleanup:
    sprintf(command,"del /F /Q %s\n",fname_tmp );
 #else
    if( do_irr_diag ) {
+      /*
+	command_argv[0] = "/bin/rm";
+	command_argv[1] = "-f";
+	command_argv[2] = fname_wrk;
+	if (posix_spawn(&child_pid, command_argv[0], NULL, NULL, command_argv, environ)) {
+          fprintf(stderr, "Could not remove %s\n", fname_wrk);
+	  exit(2);
+	}
+      */
      sprintf(command,"/bin/rm -f %s\n",fname_wrk );
      system( command ) ;
    }
